+++ inst/NEWS.md
@@ -15,7 +15,18 @@
 
 * Fix color detection in Emacs tramp
 
-* Some performance improvements to `col_substr`
+* `col_strsplit` and `col_substr` corner cases:
+
+    * handle empty chunks at beginning or end of strings
+      like `base::strsplit` (@brodieG, #26)
+
+    * explicitly deal with 'split' values that are not
+      length 1 as that is not currently supported
+
+    * handle zero length `x` argument in `col_substr`, and
+      add more explicit error messages for corner cases
+
+* Some performance improvements to `col_substr` (@brodieG)
 
 * Change rgb to ANSI code mapping, based on the "paint" ruby gem
   (@richfitz, #33, #34)